Extract calculatePeriodDays to utils
