+++ src/lib/utils.js
@@ -5,6 +5,17 @@ export function cn(...inputs) {
   return twMerge(clsx(inputs))
 }
 
+export function calculatePeriodDays(startDate, endDate) {
+  // stringToDate ensures these match the day cell date in the UI
+  const start = stringToDate(startDate);
+  const end = stringToDate(endDate);
+  const days = [];
+  for (let d = start; d <= end; d.setDate(d.getDate() + 1)) {
+    days.push(formatDate(d));
+  }
+  return days;
+}
+
 // String(?) to YYYY-MM-DD
 export function formatDate(date) {
   if (!date) return '';
